# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FMD2r",
    "description": "",
    "license": {
      "name": "GPL-2.0-only",
      "identifier": "GPL-2.0-only"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/modules/{id}/settings": {
      "get": {
        "tags": [
          "modules"
        ],
        "summary": "A module's options, limits and overrides.",
        "operationId": "getModuleSettings",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Module ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The module's settings",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModuleSettingsView"
                }
              }
            }
          },
          "404": {
            "description": "No module with that ID is loaded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "modules"
        ],
        "summary": "Update a module's settings with a JSON merge patch (RFC 7396) over `enabled`, `limits`,\n`http`, `save_to` and `options` (option values keyed by `key`; `null` resets one to its default).\nNothing is stored unless the whole patch is valid.",
        "operationId": "patchModuleSettings",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Module ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": {},
                "propertyNames": {
                  "type": "string"
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "The updated settings",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModuleSettingsView"
                }
              }
            }
          },
          "400": {
            "description": "The patch is not a JSON object",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "404": {
            "description": "No module with that ID is loaded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "A value is invalid or a setting unknown; `field` names it",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "FieldProblem": {
        "type": "object",
        "description": "One rejected value of a 422.",
        "required": [
          "field",
          "detail"
        ],
        "properties": {
          "detail": {
            "type": "string"
          },
          "field": {
            "type": "string",
            "description": "The setting as a dotted path, like [`Problem::field`]."
          }
        }
      },
...
      "HttpOverridesView": {
        "type": "object",
        "description": "[`HttpOverrides`] with the proxy password left out.",
        "required": [
          "user_agent",
          "cookies",
          "proxy"
        ],
        "properties": {
          "cookies": {
            "type": "string",
            "description": "Cookies merged into every request (`HTTP.Cookies`,\nbaseunits/WebsiteModulesSettings.pas:41)."
          },
          "proxy": {
            "$ref": "#/components/schemas/ProxyOverrideView"
          },
          "user_agent": {
            "type": "string",
            "description": "Replaces the default user agent when non-empty."
          }
        }
      },
...
      "LimitOverrides": {
        "type": "object",
        "description": "Limit overrides; 0 means \"not overridden\" for tasks and threads\n(baseunits/WebsiteModulesSettings.pas:81-83).",
        "properties": {
          "max_connection_limit": {
            "type": "integer",
            "format": "int32",
            "description": "Replaces the module's connection limit while enabled, 0 (unlimited) included\n(baseunits/WebsiteModulesSettings.pas:126-155).",
            "default": 0,
            "minimum": 0
          },
          "max_task_limit": {
            "type": "integer",
            "format": "int32",
            "default": 0,
            "minimum": 0
          },
          "max_thread_per_task_limit": {
            "type": "integer",
            "format": "int32",
            "default": 0,
            "minimum": 0
          }
        }
      },
...
      "ModuleLimits": {
        "type": "object",
        "description": "The limits a module declares (`MaxTaskLimit`, `MaxThreadPerTaskLimit`, `MaxConnectionLimit`\non the Lua `MODULE` object); 0 means unlimited.",
        "required": [
          "max_task_limit",
          "max_thread_per_task_limit",
          "max_connection_limit"
        ],
        "properties": {
          "max_connection_limit": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "max_task_limit": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "max_thread_per_task_limit": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          }
        }
      },
      "ModuleOptionSetting": {
        "oneOf": [
          {
            "type": "object",
            "description": "`AddOptionCheckBox`.",
            "required": [
              "key",
              "caption",
              "default",
              "value",
              "kind"
            ],
            "properties": {
              "caption": {
                "type": "string"
              },
              "default": {
                "type": "boolean"
              },
              "key": {
                "type": "string"
              },
              "kind": {
                "type": "string",
                "enum": [
                  "checkbox"
                ]
              },
              "value": {
                "type": "boolean"
              }
            }
          },
          {
            "type": "object",
            "description": "`AddOptionEdit`.",
            "required": [
              "key",
              "caption",
              "default",
              "value",
              "kind"
            ],
            "properties": {
              "caption": {
                "type": "string"
              },
              "default": {
                "type": "string"
              },
              "key": {
                "type": "string"
              },
              "kind": {
                "type": "string",
                "enum": [
                  "edit"
                ]
              },
              "value": {
                "type": "string"
              }
            }
          },
          {
            "type": "object",
            "description": "`AddOptionSpinEdit`; FMD2's spin edit spans `min..=max`.",
            "required": [
              "key",
              "caption",
              "default",
              "value",
              "min",
              "max",
              "kind"
            ],
            "properties": {
              "caption": {
                "type": "string"
              },
              "default": {
                "type": "integer",
                "format": "int32"
              },
              "key": {
                "type": "string"
              },
              "kind": {
                "type": "string",
                "enum": [
                  "spinedit"
                ]
              },
              "max": {
                "type": "integer",
                "format": "int32"
              },
              "min": {
                "type": "integer",
                "format": "int32"
              },
              "value": {
                "type": "integer",
                "format": "int32"
              }
            }
          },
          {
            "type": "object",
            "description": "`AddOptionComboBox`; `default` and `value` index `items`.",
            "required": [
              "key",
              "caption",
              "items",
              "default",
              "value",
              "kind"
            ],
            "properties": {
              "caption": {
                "type": "string"
              },
              "default": {
                "type": "integer",
                "format": "int32"
              },
              "items": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              },
              "key": {
                "type": "string"
              },
              "kind": {
                "type": "string",
                "enum": [
                  "combobox"
                ]
              },
              "value": {
                "type": "integer",
                "format": "int32"
              }
            }
          }
        ],
        "description": "One declared option, its default and the value `MODULE.GetOption` returns for it.\n\n`key` is the name the value is stored and patched under."
      },
      "ModuleSettingsView": {
        "type": "object",
        "description": "A module's settings.",
        "required": [
          "id",
          "name",
          "enabled",
          "options",
          "limits",
          "module_limits",
          "http",
          "save_to"
        ],
        "properties": {
          "enabled": {
            "type": "boolean",
            "description": "Whether `limits` and `http` apply (`Settings.Enabled`); options always do."
          },
          "http": {
            "$ref": "#/components/schemas/HttpOverridesView"
          },
          "id": {
            "type": "string"
          },
          "limits": {
            "$ref": "#/components/schemas/LimitOverrides",
            "description": "The user's limit overrides."
          },
          "module_limits": {
            "$ref": "#/components/schemas/ModuleLimits",
            "description": "The limits the module itself declares; 0 means unlimited."
          },
          "name": {
            "type": "string"
          },
          "options": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ModuleOptionSetting"
            },
            "description": "The options the module declares with `AddOption*`, in declaration order."
          },
          "save_to": {
            "type": "string",
            "description": "The folder the website's downloads go to when the user picks none; empty for the default\ndestination (`OverrideSettings.SaveToPath`, baseunits/WebsiteModulesSettings.pas:50).\nApplies whether or not `enabled` is set."
          }
        }
      },
...
      "Problem": {
        "type": "object",
        "description": "RFC 9457 problem details body.",
        "required": [
          "type",
          "title",
          "status",
          "detail"
        ],
        "properties": {
          "detail": {
            "type": "string"
          },
          "field": {
            "type": [
              "string",
              "null"
            ],
            "description": "The setting a validation error (422) is about, as a dotted path such as\n`connections.timeout_secs` or `options.server`.\nKept for clients that read one field: the first of `fields`."
          },
          "fields": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FieldProblem"
            },
            "description": "Every setting a validation error (422) is about, each with why it was rejected."
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "title": {
            "type": "string",
            "description": "The status code's reason phrase."
          },
          "type": {
            "type": "string",
            "description": "Always `about:blank`: the status code says it all."
          }
        }
      },
      "ProxyOverrideType": {
        "type": "string",
        "description": "`TProxyType` (baseunits/WebsiteModulesSettings.pas:11).",
        "enum": [
          "default",
          "direct",
          "http",
          "socks4",
          "socks5"
        ]
      },
      "ProxyOverrideView": {
        "type": "object",
        "description": "[`ProxyOverride`] with the password left out.",
        "required": [
          "type",
          "host",
          "port",
          "username",
          "has_password"
        ],
        "properties": {
          "has_password": {
            "type": "boolean",
            "description": "Whether a password is set. Patch `password` to set it; an empty one clears it."
          },
          "host": {
            "type": "string"
          },
          "port": {
            "type": "string"
          },
          "type": {
            "$ref": "#/components/schemas/ProxyOverrideType"
          },
          "username": {
            "type": "string"
          }
        }
      },
...
    }
  }
}